Move a method to a new file
