+++ api/src/server.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"fmt"
 	"net/http"
 
 	"cloud.google.com/go/firestore"
@@ -57,62 +56,6 @@ func NewServer(c ServerConfig) *Server {
 		collectionAPIKey:     c.CollectionAPIKey,
 		analyticsIDSecret:    c.AnalitycsIDSecret,
 	}
-}
-
-// MembersWebhook accepts push notifications from Members system
-func (s Server) MembersWebhook(c *gin.Context) {
-	msg, err := pubsub.MessageFromCtx(c)
-	if err != nil {
-		data, _ := c.GetRawData()
-		log.L.Info().
-			Str("body", string(data)).
-			Err(err).
-			Msg("Malformed request")
-		c.JSON(http.StatusNoContent, map[string]string{"message": "Malformed request 1"})
-		return
-	}
-
-	if !msg.Validate(s.membersWebhookSecret) {
-		log.L.Debug().
-			Str("data", msg.Message.Data).
-			Str("msg", fmt.Sprintf("%+v", msg)).
-			Msg("Invalid HMAC")
-		c.JSON(http.StatusOK, gin.H{"error": "Message not signed correctly"})
-		return
-	}
-
-	updatedMemebers := []members.Member{}
-	err = msg.ExtractDataInto(&updatedMemebers)
-	if err != nil {
-		log.L.Info().
-			Err(err).
-			Str("data", msg.Message.Data).
-			Msg("Malformed request")
-		c.JSON(http.StatusOK, map[string]string{"message": "Malformed request 2"})
-		return
-	}
-
-	for i, person := range updatedMemebers {
-		err = firebase.UpdateOrCreateUser(c.Request.Context(), s.fs, &person)
-		if err == nil {
-			log.L.Debug().
-				Int("person_id", person.PersonID).
-				Msg("Updated person")
-			continue
-		}
-
-		log.L.Error().
-			Err(err).
-			Int("person_id", person.PersonID).
-			Int("i", i).
-			Msg("Error when updating user")
-
-		// We just log the errors, for now as it it unlikely that we will be able to do
-		// better in a retry
-	}
-
-	// We have to return a 2xx code regardless of the actual success as we don't want to get the message again
-	c.Status(http.StatusNoContent)
 }
 
 // UpdatePersonRequest for parsing the PubSub message
